backend/requirements.txt
fastapi
uvicorn
pydantic
google-cloud-firestore
firebase-admin
python-dotenv
requests
google-generativeai 